requirements.txt
openai
langchain
sqlalchemy
pyodbc
streamlit
python-dotenv
